get rid of dedicated dask directory
