+++ tests/unit/test_workflow.py
@@ -24,8 +24,8 @@ from cudf.tests.utils import assert_eq
 from pandas.api.types import is_integer_dtype
 
 import nvtabular as nvt
-import nvtabular.io
 import nvtabular.ops as ops
+from nvtabular.io import GPUDatasetIterator, GPUFileIterator
 from tests.conftest import allcols_csv, get_cats, mycols_csv, mycols_pq
 
 
@@ -89,7 +89,7 @@ def test_gpu_workflow_api(
     # Write to new "shuffled" and "processed" dataset
     processor.write_to_dataset(tmpdir, dataset, nfiles=10, shuffle=True, apply_ops=True)
 
-    data_itr_2 = nvtabular.io.GPUDatasetIterator(
+    data_itr_2 = GPUDatasetIterator(
         glob.glob(str(tmpdir) + "/ds_part.*.parquet"),
         use_row_groups=True,
         gpu_memory_frac=gpu_memory_frac,
@@ -112,9 +112,7 @@ def test_gpu_file_iterator_parquet(datasets, batch):
     paths = glob.glob(str(datasets["parquet"]) + "/*.parquet")
     df_expect = cudf.read_parquet(paths[0], columns=mycols_pq)
     df_itr = cudf.DataFrame()
-    data_itr = nvtabular.io.GPUFileIterator(
-        paths[0], batch_size=batch, gpu_memory_frac=0.01, columns=mycols_pq
-    )
+    data_itr = GPUFileIterator(paths[0], batch_size=batch, gpu_memory_frac=0.01, columns=mycols_pq)
     for data_gd in data_itr:
         df_itr = cudf.concat([df_itr, data_gd], axis=0) if df_itr else data_gd
 
@@ -130,7 +128,7 @@ def test_gpu_file_iterator_csv(datasets, batch, dskey):
     df_expect = cudf.read_csv(paths[0], header=header, names=names)[mycols_csv]
     df_expect["id"] = df_expect["id"].astype("int64")
     df_itr = cudf.DataFrame()
-    data_itr = nvtabular.io.GPUFileIterator(
+    data_itr = GPUFileIterator(
         paths[0], batch_size=batch, gpu_memory_frac=0.01, columns=mycols_csv, names=names
     )
     for data_gd in data_itr:
@@ -145,9 +143,7 @@ def test_gpu_dataset_iterator_parquet(datasets, batch):
     df_expect = cudf.read_parquet(paths[0], columns=mycols_pq)
     df_expect = cudf.concat([df_expect, cudf.read_parquet(paths[1], columns=mycols_pq)], axis=0)
     df_itr = cudf.DataFrame()
-    data_itr = nvtabular.io.GPUDatasetIterator(
-        paths, batch_size=batch, gpu_memory_frac=0.01, columns=mycols_pq
-    )
+    data_itr = GPUDatasetIterator(paths, batch_size=batch, gpu_memory_frac=0.01, columns=mycols_pq)
     for data_gd in data_itr:
         df_itr = cudf.concat([df_itr, data_gd], axis=0) if df_itr else data_gd
 
@@ -216,7 +212,7 @@ def test_gpu_workflow(tmpdir, client, df, dataset, gpu_memory_frac, engine, dump
     # Write to new "shuffled" and "processed" dataset
     processor.write_to_dataset(tmpdir, dataset, nfiles=10, shuffle=True, apply_ops=True)
 
-    data_itr_2 = nvtabular.io.GPUDatasetIterator(
+    data_itr_2 = GPUDatasetIterator(
         glob.glob(str(tmpdir) + "/ds_part.*.parquet"),
         use_row_groups=True,
         gpu_memory_frac=gpu_memory_frac,
@@ -304,7 +300,7 @@ def test_gpu_workflow_config(tmpdir, client, df, dataset, gpu_memory_frac, engin
     # Write to new "shuffled" and "processed" dataset
     processor.write_to_dataset(tmpdir, dataset, nfiles=10, shuffle=True, apply_ops=True)
 
-    data_itr_2 = nvtabular.io.GPUDatasetIterator(
+    data_itr_2 = GPUDatasetIterator(
         glob.glob(str(tmpdir) + "/ds_part.*.parquet"),
         use_row_groups=True,
         gpu_memory_frac=gpu_memory_frac,
